Factor out interface-specific early_adopter code
Refactors early_adopter such that interface-specific code is located in
the framework module as a submodule `alpha`.
