+++ src/python/src/grpc/framework/alpha/__init__.py
@@ -26,61 +26,3 @@
 # THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
 # (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 # OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-
-"""A setup module for the GRPC Python package."""
-
-from distutils import core as _core
-
-_EXTENSION_SOURCES = (
-    'grpc/_adapter/_c.c',
-    'grpc/_adapter/_call.c',
-    'grpc/_adapter/_channel.c',
-    'grpc/_adapter/_completion_queue.c',
-    'grpc/_adapter/_error.c',
-    'grpc/_adapter/_server.c',
-    'grpc/_adapter/_client_credentials.c',
-    'grpc/_adapter/_server_credentials.c',
-)
-
-_EXTENSION_INCLUDE_DIRECTORIES = (
-    '.',
-)
-
-_EXTENSION_LIBRARIES = (
-    'grpc',
-    'gpr',
-    'rt',
-)
-
-_EXTENSION_MODULE = _core.Extension(
-    'grpc._adapter._c', sources=list(_EXTENSION_SOURCES),
-    include_dirs=list(_EXTENSION_INCLUDE_DIRECTORIES),
-    libraries=list(_EXTENSION_LIBRARIES),
-    )
-
-_PACKAGES = (
-    'grpc',
-    'grpc._adapter',
-    'grpc._junkdrawer',
-    'grpc.early_adopter',
-    'grpc.framework',
-    'grpc.framework.alpha',
-    'grpc.framework.base',
-    'grpc.framework.common',
-    'grpc.framework.face',
-    'grpc.framework.face.testing',
-    'grpc.framework.foundation',
-)
-
-_PACKAGE_DIRECTORIES = {
-    'grpc': 'grpc',
-    'grpc._adapter': 'grpc/_adapter',
-    'grpc._junkdrawer': 'grpc/_junkdrawer',
-    'grpc.early_adopter': 'grpc/early_adopter',
-    'grpc.framework': 'grpc/framework',
-}
-
-_core.setup(
-    name='grpc-2015', version='0.4.0',
-    ext_modules=[_EXTENSION_MODULE], packages=list(_PACKAGES),
-    package_dir=_PACKAGE_DIRECTORIES)
