+++ src/python/src/grpc/framework/alpha/_reexport.py
@@ -31,8 +31,8 @@ from grpc.framework.common import cardinality
 from grpc.framework.face import exceptions as face_exceptions
 from grpc.framework.face import interfaces as face_interfaces
 from grpc.framework.foundation import future
-from grpc.early_adopter import exceptions
-from grpc.early_adopter import interfaces
+from grpc.framework.alpha import exceptions
+from grpc.framework.alpha import interfaces
 
 _EARLY_ADOPTER_CARDINALITY_TO_COMMON_CARDINALITY = {
     interfaces.Cardinality.UNARY_UNARY: cardinality.Cardinality.UNARY_UNARY,
